Make use of reusable thread-bound buffers instead of allocating a new ByteArrayOutputStream for every log event
