+++ src/main/java/net/logstash/logback/composite/CompositeJsonFormatter.java
@@ -16,28 +16,27 @@ package net.logstash.logback.composite;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.io.Writer;
-import java.lang.ref.SoftReference;
 import java.util.ServiceConfigurationError;
 
 import net.logstash.logback.decorate.JsonFactoryDecorator;
 import net.logstash.logback.decorate.JsonGeneratorDecorator;
 import net.logstash.logback.decorate.NullJsonFactoryDecorator;
 import net.logstash.logback.decorate.NullJsonGeneratorDecorator;
+import net.logstash.logback.util.ThreadLocalBuffers;
+
+import com.fasterxml.jackson.core.JsonEncoding;
+import com.fasterxml.jackson.core.JsonFactory;
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.core.io.SegmentedStringWriter;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
+
 import ch.qos.logback.access.spi.IAccessEvent;
 import ch.qos.logback.classic.spi.ILoggingEvent;
 import ch.qos.logback.core.spi.ContextAware;
 import ch.qos.logback.core.spi.ContextAwareBase;
 import ch.qos.logback.core.spi.DeferredProcessingAware;
 import ch.qos.logback.core.spi.LifeCycle;
-
-import com.fasterxml.jackson.core.JsonEncoding;
-import com.fasterxml.jackson.core.JsonFactory;
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.core.io.SegmentedStringWriter;
-import com.fasterxml.jackson.core.util.BufferRecycler;
-import com.fasterxml.jackson.core.util.ByteArrayBuilder;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
 
 /**
  * Formats logstash Events as JSON using {@link JsonProvider}s.
@@ -53,16 +52,10 @@ public abstract class CompositeJsonFormatter<Event extends DeferredProcessingAwa
         extends ContextAwareBase implements LifeCycle {
 
     /**
-     * This <code>ThreadLocal</code> contains a {@link java.lang.ref.SoftReference}
-     * to a {@link BufferRecycler} used to provide a low-cost
-     * buffer recycling between writer instances.
+     * Provides thread local buffers
      */
-    private final ThreadLocal<SoftReference<BufferRecycler>> recycler = new ThreadLocal<SoftReference<BufferRecycler>>() {
-        protected SoftReference<BufferRecycler> initialValue() {
-            final BufferRecycler bufferRecycler = new BufferRecycler();
-            return new SoftReference<BufferRecycler>(bufferRecycler);
-        }
-    };
+    private final ThreadLocalBuffers buffers = new ThreadLocalBuffers();
+    
     
     /**
      * Used to create the necessary {@link JsonGenerator}s for generating JSON.
@@ -153,18 +146,6 @@ public abstract class CompositeJsonFormatter<Event extends DeferredProcessingAwa
         return this.jsonFactoryDecorator.decorate(jsonFactory);
     }
 
-    public byte[] writeEventAsBytes(Event event) throws IOException {
-        ByteArrayBuilder outputStream = new ByteArrayBuilder(getBufferRecycler());
-
-        try {
-            writeEventToOutputStream(event, outputStream);
-            outputStream.flush();
-            return outputStream.toByteArray();
-        } finally {
-            outputStream.release();
-        }
-    }
-
     public void writeEventToOutputStream(Event event, OutputStream outputStream) throws IOException {
         try (JsonGenerator generator = createGenerator(outputStream)) {
             writeEventToGenerator(generator, event);
@@ -178,12 +159,15 @@ public abstract class CompositeJsonFormatter<Event extends DeferredProcessingAwa
     }
 
     public String writeEventAsString(Event event) throws IOException {
-        SegmentedStringWriter writer = new SegmentedStringWriter(getBufferRecycler());
+        SegmentedStringWriter writer = this.buffers.getStringWriter();
 
         try (JsonGenerator generator = createGenerator(writer)) {
             writeEventToGenerator(generator, event);
-            writer.flush();
             return writer.getAndClear();
+        }
+        finally {
+            // make sure the buffer is released!
+            writer.getAndClear();
         }
     }
 
@@ -208,16 +192,6 @@ public abstract class CompositeJsonFormatter<Event extends DeferredProcessingAwa
 
     private JsonGenerator createGenerator(Writer writer) throws IOException {
         return this.jsonGeneratorDecorator.decorate(jsonFactory.createGenerator(writer));
-    }
-
-    private BufferRecycler getBufferRecycler() {
-        SoftReference<BufferRecycler> bufferRecyclerReference = recycler.get();
-        BufferRecycler bufferRecycler = bufferRecyclerReference.get();
-        if (bufferRecycler == null) {
-            recycler.remove();
-            return getBufferRecycler();
-        }
-        return bufferRecycler;
     }
     
     public JsonFactory getJsonFactory() {
